requirements.txt
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.6
openai>=1.3.0
chromadb>=0.4.15
tiktoken>=0.5.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
requests>=2.28.0
streamlit
